fix survey querier
