+++ backend/src/main/java/com/javahelp/backend/data/search/constraint/Constraint.java
@@ -3,21 +3,25 @@ package com.javahelp.backend.data.search.constraint;
 import java.util.HashSet;
 import java.util.Set;
 
+/**
+ * Constraint on a query for {@link com.javahelp.model.survey.SurveyResponse}s
+ */
 public class Constraint implements IConstraint {
-    private Set<String> constraints = new HashSet<>();
 
-    public Constraint() {}
+    private final Set<String> constraints = new HashSet<>();
 
     @Override
-    public void setConstraint(String constraint) {
-        this.constraints.add(constraint);
+    public void addConstraint(String constraint) {
+        constraints.add(constraint);
     }
 
     @Override
-    public Set<String> getConstraints(){
-        return this.constraints;
-    };
+    public Set<String> getConstraints() {
+        return constraints;
+    }
 
     @Override
-    public int size() { return this.constraints.size(); }
+    public int size() {
+        return constraints.size();
+    }
 }
